requirements.txt
# Core dependencies
requests==2.32.3
python-dotenv==1.0.1
openai==1.32.0

# PDF processing (для работы с базой знаний)
PyMuPDF==1.24.1

# Vector search (для File Search)
faiss-cpu==1.8.0

# Telegram bot (для будущей версии)
python-telegram-bot==20.7

# Testing
pytest==7.4.3

